Inter-map movement first implementation
Inter-map movement functional with Car as a ground type and an Exit
added at its location. Will try to refactor for Car as an item, to
reduce possibility for bugs as a result of Car location and Exit
location mismatch, and reduce repetition if more cars required to be
added.