Further updates to Sniper Rifle implementation
Shooting on the first turn fully implemented, Aiming for subsequent
shooting partially implemented. FireSniperAction complete, SniperRifle
just needs a proper switch to feed the right information into the firing
action.

Refactored so that aiming status is no longer stored in the Player but
in the rifle; this is to avoid casting down and so violating LSP. Will
need to work on another way to clear aimingstatus when another action is
taken.